separated models from neural ucb algorithm
