+++ metric_bandits/nets/README.md
@@ -1,0 +1,1 @@
+Contains the implementations of various nets that can be used alongside neural-ucb
